remove: remove social login (due to cognito spec)
