--- src/app/api/auth/signin/google/route.ts
@@ -1,3 +1,0 @@
-export async function POST(request: Request) {
-  const token = request.headers.get("authorization");
-}
