--- src/app/auth/link/google/page.tsx
@@ -1,13 +1,0 @@
-"use client";
-
-import dynamic from "next/dynamic";
-
-const GoogleAccountLink = dynamic(() => import("./GoogleAccountLInk"), { ssr: false });
-
-export default function AuthWithGoogle() {
-  return (
-    <>
-      <GoogleAccountLink />
-    </>
-  );
-}
