--- src/app/auth/signup/google/page.tsx
@@ -1,13 +1,0 @@
-"use client";
-
-import dynamic from "next/dynamic";
-
-const GetToken = dynamic(() => import("./SocialSignupForm"), { ssr: false });
-
-export default function AuthWithGoogle() {
-  return (
-    <>
-      <GetToken />
-    </>
-  );
-}
