+++ src/app/auth/signup/page.tsx
@@ -1,30 +1,102 @@
-import SocialButton from "components/SocialButton";
+"use client";
+
 import Link from "next/link";
+import { useState } from "react";
+import { useRouter } from "next/navigation";
 
-// https://<YOUR_DOMAIN>.auth.<REGION>.amazoncognito.com/oauth2/authorize?identity_provider=Google&redirect_uri=<REDIRECT_URI>&response_type=TOKEN&client_id=<CLIENT_ID>&scope=openid
+export default function AuthWithEmail() {
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [nickname, setNickname] = useState("");
+  const router = useRouter();
 
-export default function SignUpPage() {
+  const onChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const {
+      target: { name, value },
+    } = event;
+    const excuteFunction =
+      name === "email" ? setEmail : name === "password" ? setPassword : name === "nickname" ? setNickname : () => {};
+    excuteFunction(value);
+  };
+
+  const onSubmit = async (event: React.FormEvent) => {
+    event.preventDefault();
+    const userData = {
+      email: email,
+      password: password,
+      nickname: nickname,
+    };
+    try {
+      const result = await fetch(`${process.env.NEXT_PUBLIC_API_DOMAIN}/auth/signup`, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(userData),
+      });
+      if (!result.ok) {
+        const { error } = await result.json();
+        throw new Error(error);
+      } else router.push(`/auth/verification?email=${email}`);
+    } catch (error) {
+      console.log(error);
+      window.alert("Sign up Error");
+    }
+  };
+
   return (
-    <>
-      <div className="AuthWithSocialAccount vstack gap-3">
-        <SocialButton
-          name="Google"
-          img={"/google.png"}
-          href={
-            `${process.env.NEXT_PUBLIC_COGNITO_DOMAIN}/oauth2/authorize?` +
-            `identity_provider=Google&` +
-            `redirect_uri=${process.env.NEXT_PUBLIC_WEB_DOMAIN}/auth/signup/google/&` +
-            "response_type=TOKEN&" +
-            `client_id=${process.env.COGNITO_CLIENT_ID}&` +
-            "scope=email+openid+profile"
-          }
-        />
-        <SocialButton name="Facebook" img={"/facebook.png"} href={"/auth/signup/facebook"} />
-        <SocialButton name="Email" img={"/email.png"} href={"/auth/signup/email"} />
-      </div>
-      <Link className="btn btn-primary col-8 offset-2 h-36px" href={"/auth/signin"} role="button">
-        {"Sign in"}
-      </Link>
-    </>
+    <div className="AuthWithEmail vstack gap-3">
+      <form onSubmit={onSubmit}>
+        <div className="vstack gap-3">
+          <div>
+            <label className="form-label">Email address</label>
+            <input
+              className="form-control"
+              name="email"
+              type="text"
+              placeholder="email"
+              value={email}
+              required
+              onChange={onChange}
+            />
+          </div>
+          <div>
+            <label className="form-label">Password</label>
+            <input
+              className="form-control"
+              name="password"
+              type="password"
+              placeholder="password"
+              value={password}
+              autoComplete="off"
+              required
+              onChange={onChange}
+            />
+          </div>
+          <div>
+            <label className="form-label">Nickname</label>
+            <input
+              className="form-control"
+              name="nickname"
+              type="text"
+              placeholder="4-20 digits of English, numbers and special characters"
+              value={nickname}
+              autoComplete="off"
+              required
+              onChange={onChange}
+            />
+          </div>
+          <button type="submit" className="btn btn-primary col-8 offset-2 h-36px">
+            {"Create Account"}
+          </button>
+          <Link className="btn btn-info col-8 offset-2 h-36px" href="/auth/signin" role="button">
+            {"Do you have an account?"}
+          </Link>
+          <Link className="btn btn-secondary col-8 offset-2 h-36px" href="/" role="button">
+            {"Back"}
+          </Link>
+        </div>
+      </form>
+    </div>
   );
 }
