+++ src/app/setting/page.tsx
@@ -4,7 +4,6 @@ import getCurrentUserData from "logics/getCurrentUserData";
 import SettingData from "./SettingData";
 import ProfileInfoEdit from "./ProfileInfoEdit";
 import ProfileImageEdit from "./ProfileImageEdit";
-import SocialLoginEdit from "./SocialLoginEdit";
 
 export default function SettingPage() {
   const { data: userData, status } = useQuery({
@@ -44,7 +43,6 @@ export default function SettingPage() {
             currentData={userData?.email}
             onClick={onEmailChange}
           />
-          {<SocialLoginEdit />}
           {
             // <CategoryOrderEdit />
           }
